pack: production

See merge request castlecraft/excel-rma!1029

+++ packages/rma-server/src/stock-entry/aggregates/warranty-stock-entry-aggregate/warranty-stock-entry-aggregate.service.ts
@@ -8,8 +8,10 @@ import {
 import { StockEntryService } from '../../entities/stock-entry.service';
 import { from, throwError, of, forkJoin } from 'rxjs';
 import {
+  CURRENT_STATUS_VERDICT,
   DEFAULT_NAMING_SERIES,
   DELIVERY_STATUS,
+  NON_SERIAL_ITEM,
   STOCK_ENTRY,
   STOCK_ENTRY_STATUS,
   VERDICT,
@@ -33,10 +35,7 @@ import {
 } from '../../../serial-no/entity/serial-no-history/serial-no-history.entity';
 import { StockEntryPoliciesService } from '../../../stock-entry/policies/stock-entry-policies/stock-entry-policies.service';
 import { WarrantyClaimAggregateService } from '../../../warranty-claim/aggregates/warranty-claim-aggregate/warranty-claim-aggregate.service';
-import {
-  INVOICE_DELIVERY_STATUS,
-  PROGRESS_STATUS,
-} from '../../../constants/app-strings';
+import { PROGRESS_STATUS } from '../../../constants/app-strings';
 
 @Injectable()
 export class WarrantyStockEntryAggregateService {
@@ -335,7 +334,7 @@ export class WarrantyStockEntryAggregateService {
 
   updateSerials(deliveryNote, serial_no, settings) {
     if (deliveryNote.isSync) {
-      return of();
+      return of(true);
     }
     if (
       deliveryNote.items[0].excel_serials &&
@@ -444,8 +443,9 @@ export class WarrantyStockEntryAggregateService {
             {
               $set: {
                 customer: state.progress_state[0].customer,
-                'warranty.salesWarrantyDate':
-                  state.progress_state[0].salesWarrantyDate,
+                'warranty.salesWarrantyDate': state.warranty_end_date
+                  ? state.warranty_end_date
+                  : '',
                 'warranty.soldOn': new DateTime(settings.timeZone).toJSDate(),
                 sales_invoice_name: state.progress_state[0].sales_invoice_name,
                 delivery_note: payload.delivery_note,
@@ -506,7 +506,7 @@ export class WarrantyStockEntryAggregateService {
     return from(this.stockEntryService.findOne(warrantyClaimUuid));
   }
 
-  removeStockEntry(stockEntry, req) {
+  removeStockEntry(stockEntry: WarrantyStockEntryDto, req) {
     let set: any;
     return this.stockEntryPoliciesService
       .validateCancelWarrantyStockEntry(stockEntry.warrantyClaimUuid, [
@@ -531,39 +531,52 @@ export class WarrantyStockEntryAggregateService {
           if (!response) {
             return throwError(new NotFoundException());
           }
-          const url = `${set.authServerURL}${POST_DELIVERY_NOTE_ENDPOINT}/${stockEntry.stock_voucher_number}`;
-          response.docstatus = 2;
-          response.items.forEach(item => {
-            item.docstatus = 2;
-          });
-          return this.http.put(url, response, {
-            headers: this.settingService.getAuthorizationHeaders(req.token),
-          });
+          if (response.docstatus !== 2) {
+            const url = `${set.authServerURL}${POST_DELIVERY_NOTE_ENDPOINT}/${stockEntry.stock_voucher_number}`;
+            response.docstatus = 2;
+            response.items.forEach(item => {
+              item.docstatus = 2;
+            });
+            return this.http.put(url, response, {
+              headers: this.settingService.getAuthorizationHeaders(req.token),
+            });
+          }
+          return of({ data: {} });
         }),
         map(res => res.data.data),
         switchMap(serialItem => {
-          if (stockEntry.stock_entry_type === STOCK_ENTRY_STATUS.delivered) {
-            return from(
-              this.serialService.updateOne(
-                { serial_no: stockEntry.items[0]?.serial_no },
-                {
-                  $unset: {
-                    customer: '',
-                    'warranty.salesWarrantyDate': '',
-                    'warranty.soldOn': '',
-                    delivery_note: '',
-                    sales_invoice_name: '',
+          if (
+            stockEntry.items.find(item => {
+              if (item.serial_no.toUpperCase() === NON_SERIAL_ITEM) {
+                return undefined;
+              }
+              return item.serial_no;
+            })
+          ) {
+            if (stockEntry.stock_entry_type === STOCK_ENTRY_STATUS.delivered) {
+              return from(
+                this.serialService.updateOne(
+                  { serial_no: stockEntry.items[0]?.serial_no },
+                  {
+                    $unset: {
+                      customer: '',
+                      'warranty.salesWarrantyDate': '',
+                      'warranty.soldOn': '',
+                      delivery_note: '',
+                      sales_invoice_name: '',
+                    },
                   },
-                },
-              ),
-            );
+                ),
+              );
+            }
+            if (stockEntry.stock_entry_type === STOCK_ENTRY_STATUS.returned) {
+              return this.resetCancelledSerialItem(
+                stockEntry.stock_voucher_number,
+              );
+            }
+            return of(true);
           }
-          if (stockEntry.stock_entry_type === STOCK_ENTRY_STATUS.returned) {
-            return this.resetCancelledSerialItem(
-              stockEntry.stock_voucher_number,
-            );
-          }
-          return of();
+          return of(true);
         }),
         switchMap(() => {
           if (stockEntry.stock_entry_type === STOCK_ENTRY_STATUS.delivered) {
@@ -585,7 +598,7 @@ export class WarrantyStockEntryAggregateService {
         switchMap(() => {
           return from(
             this.stockEntryService.deleteOne({
-              stock_voucher_number: stockEntry.stock_voucher_number,
+              uuid: stockEntry.uuid,
             }),
           );
         }),
@@ -610,11 +623,21 @@ export class WarrantyStockEntryAggregateService {
           );
         }),
         switchMap(() => {
-          return from(
-            this.serialNoHistoryService.deleteOne({
-              document_no: stockEntry.stock_voucher_number,
-            }),
-          );
+          if (
+            stockEntry.items.find(item => {
+              if (item.serial_no.toUpperCase() === NON_SERIAL_ITEM) {
+                return undefined;
+              }
+              return item.serial_no;
+            })
+          ) {
+            return from(
+              this.serialNoHistoryService.deleteOne({
+                document_no: stockEntry.stock_voucher_number,
+              }),
+            );
+          }
+          return of(true);
         }),
       );
   }
@@ -625,7 +648,7 @@ export class WarrantyStockEntryAggregateService {
         uuid,
         status_history: {
           $elemMatch: {
-            verdict: INVOICE_DELIVERY_STATUS.DELIVERED_TO_CUSTOMER,
+            verdict: CURRENT_STATUS_VERDICT.DELIVER_TO_CUSTOMER,
           },
         },
       }),
@@ -671,7 +694,20 @@ export class WarrantyStockEntryAggregateService {
             ),
           );
         }
-        return of({});
+        return from(
+          this.serialService.updateOne(
+            { serial_no: stockEntryObject.items[0]?.serial_no },
+            {
+              $set: {
+                'warranty.salesWarrantyDate': warranty.received_on,
+                'warranty.soldOn': warranty.received_on,
+              },
+              $unset: {
+                delivery_note: '',
+              },
+            },
+          ),
+        );
       }),
     );
   }
